refactor: segregate errors, config file not found interface
